% Generated by build script
\name{hf_datasets}
\alias{hf_datasets}
\title{Search HuggingFace datasets}
\usage{
hf_datasets()
}
\arguments{
\item{search}{Search query}
\item{author}{Filter by author/organization}
\item{limit}{Max results (default 30)}
}
\value{tibble: id, author, sha, downloads, likes, last_modified}
\description{Search HuggingFace datasets}
