% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hf.R
\name{hf_datasets}
\alias{hf_datasets}
\title{Search HuggingFace datasets}
\usage{
hf_datasets(search = NULL, author = NULL, limit = 30)
}
\arguments{
\item{search}{Search query}

\item{author}{Filter by author/organization}

\item{limit}{Max results (default 30)}
}
\value{
tibble: id, author, sha, downloads, likes, last_modified
}
\description{
Search HuggingFace datasets
}
